Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -5,10 +5,10 @@ char* welcome() {
 }
 
 int main() {
-    int a;
-    int b;
-    int c;
-    scanf("%d %d %d",&a,&b,&c);
+    float a;
+    float b;
+    float c;
+    scanf("%f %f %f",&a,&b,&c);
     printf("Average: %.2f",(a+b+c)/3);
     return 0;
 }
